requirement.txt
streamlit==1.10.0
tensorflow==2.9.1
pillow==9.1.1
numpy==1.22.4
opencv-python==4.5.5.64